+++ NEWS.md
@@ -1,3 +1,10 @@
+# healthieR 0.0.2
+
+* Added support for extracting personal information like date of birth and
+  sex from Apple Health (#1).
+  
+  
+
 # healthieR 0.0.1
 
 * Added package infastructure.